pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raga"
dynamic = ["version"]
description = "A fast, distraction-free photo viewer."
readme = { file = "LONG_DESCRIPTION.md", content-type = "text/markdown" }
license = { text = "MIT" }
authors = [
    { name = "Benevant Mathew", email = "benevantmathewv@gmail.com" }
]
requires-python = ">=3.7"

dependencies = [
    "pandas",
    "pillow",
    "pillow_heif"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]

[project.scripts]
raga = "raga.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["raga", "raga.*"]

[tool.setuptools.package-data]
"raga" = ["config/extensions.json"]

[tool.setuptools.dynamic]
version = { attr = "raga.version.__version__" }

[project.urls]
Homepage = "https://github.com/benevantmathew/raga"
